[Design] #1 버튼높이에 쓸 수 있게 CGFloat으로 변경

+++ MogakStudyUIFramework/msButtons.swift
@@ -15,7 +15,7 @@ enum ButtonStatus {
     case disable
 }
 
-enum ButtonHeight: Int {
+enum ButtonHeight: CGFloat {
     case h48 = 48
     case h40 = 40
     case h32_Icon = 32
